Post generation and edit update

+++ src/helper/openai.js
@@ -44,22 +44,24 @@ OPEN_AI.generatePostVariation = async (data) => {
             try {
 
                 const resizedImageBuffer = fs.createReadStream(data.image);
-                // let img = await openai.images.createVariation({
-                //     image: resizedImageBuffer,
-                //     n: 1,
-                //     size: "1024x1024"
-                // });
-
-                let img = await openai.images.edit({
-                    // model: "dall-e-3",
+                let img = await openai.images.createVariation({
                     image: resizedImageBuffer,
-                    prompt: 'Generate a variation of this image',
                     n: 1,
                     size: "1024x1024"
                 });
+
+                // let img = await openai.images.edit({
+                //     // model: "dall-e-3",
+                //     image: resizedImageBuffer,
+                //     prompt: 'Generate a variation of this image',
+                //     n: 1,
+                //     size: "1024x1024"
+                // });
                 console.log('img', img);
 
+                fs.unlinkSync(data.image);
                 if (img.data[0].url) {
+
                     response.image = img.data[0].url
                 }
                 else {
@@ -68,8 +70,8 @@ OPEN_AI.generatePostVariation = async (data) => {
             }
             catch (err) {
                 console.log('err image', err);
-                reject(err)
                 error = true
+                reject(err)
             }
         }
         if (error) {
@@ -82,46 +84,53 @@ OPEN_AI.generatePostVariation = async (data) => {
 }
 
 OPEN_AI.generateImageUsingPrompt = async (type, prompt, file) => {
-    try {
-        let response
-
-        const resizedImagePath = "public/images/resized-" + `${Date.now()}.png`;
+    return new Promise(async (resolve, reject) => {
         try {
-            await sharp(file)
-                .resize({ width: 1024, height: 1024 })
-                .toFormat('png')
-                .ensureAlpha()
-                .toFile(resizedImagePath);
+            let response
+            const resizedImagePath = "public/images/resized-" + `${Date.now()}.png`;
+            try {
+                await sharp(file)
+                    .resize({ width: 1024, height: 1024 })
+                    .toFormat('png')
+                    .ensureAlpha()
+                    .toFile(resizedImagePath);
 
-            // console.log('Image resized successfully:', resizedImagePath);
+                // console.log('Image resized successfully:', resizedImagePath);
+            } catch (error) {
+                console.error('Error resizing image:', error);
+                return;
+            }
+
+            const resizedImageBuffer = fs.createReadStream(resizedImagePath);
+
+            // response = await openai.images.edit({
+            //     // model: "dall-e-2",
+            //     image: resizedImageBuffer,
+            //     prompt: prompt,
+            //     n: 1,
+            //     size: "1024x1024"
+            // });
+            response = await openai.images.createVariation({
+                model: "dall-e-2",
+                image: resizedImageBuffer,
+                n: 1,
+                size: "1024x1024"
+            });
+
+            if (response) {
+                fs.unlinkSync(file);
+                fs.unlinkSync(resizedImagePath);
+                let image_url = response.data[0].url;
+                resolve(image_url)
+            }
+            else {
+                reject('Failed to generate image')
+            }
         } catch (error) {
-            console.error('Error resizing image:', error);
-            return;
+            console.error(`Error generating image: ${error}`);
+            reject(error)
         }
-
-        const resizedImageBuffer = fs.createReadStream(resizedImagePath);
-
-        response = await openai.images.edit({
-            // model: "dall-e-2",
-            image: resizedImageBuffer,
-            prompt: prompt,
-            n: 1,
-            size: "1024x1024"
-        });
-        // response = await openai.images.createVariation({
-        //     model: "dall-e-2",
-        //     image: resizedImageBuffer,
-        //     n: 1,
-        //     size: "1024x1024"
-        // });
-
-        fs.unlinkSync(file);
-        fs.unlinkSync(resizedImagePath);
-        let image_url = response.data[0].url;
-        return image_url
-    } catch (error) {
-        console.error(`Error generating image: ${error}`);
-    }
+    })
 }
 
 OPEN_AI.generateDataUsingOpenAI = async (data) => {
@@ -133,23 +142,27 @@ OPEN_AI.generateDataUsingOpenAI = async (data) => {
 content : ${data.captionPrompt}
 word length : ${data.approxWords}
 tone - ${data.toneofVoice}
-${data.generateHashtags ? 'Generate Hashtags' : ''}
+${data.generateHashtags ? 'Also generate hashtags which are suitable for my content.' : ''}
 
-Generate a variation of my content while maintaining the same word length and appropriate tone. Output will be only generated variation. Make sure to not add any other content.`
+Generate a variation of given content while maintaining the same word length and appropriate tone. Output will be only generated variation. Make sure to not add any other content.`
 
 
-        const completion = await openai.chat.completions.create({
-            model: "gpt-3.5-turbo",
+        await openai.chat.completions.create({
+            model: "gpt-4o",
             messages: [{ role: "system", content: `You are an expert in analyzing and generating variation of given content. Your expertise includes understanding the nuances of the content, identifying key elements, and creating effective, relevant variation that align with the context and purpose.` },
             { role: 'user', content: content }
             ],
-        })
-        if (completion) {
-            resolve(completion.choices[0].message.content)
-        }
-        else {
+        }).then(completion => {
+            if (completion) {
+                resolve(completion.choices[0].message.content)
+            }
+            else {
+                reject('Failed to generate text. Try again!')
+            }
+        }).catch(err => {
             reject('Failed to generate text. Try again!')
-        }
+        })
+
     })
 }
 module.exports = OPEN_AI;
